backend/requirements.txt
django-cors-headers>=3.2.1,<3.3
django>=2.2,<2.3
djangorestframework>=3.11.0,<3.12
djangorestframework-simplejwt>=4.4.0,<4.5